+++ core/src/main/kotlin/fr/xgouchet/elmyr/ForgeConfigurator.kt
@@ -9,4 +9,11 @@ interface ForgeConfigurator {
      * Allows you to configure a [Forge] instance.
      */
     fun configure(forge: Forge)
+
+    /**
+     * A no-op implementation of [ForgeConfigurator].
+     */
+    object NoOp : ForgeConfigurator {
+        override fun configure(forge: Forge) {}
+    }
 }