+++ junit5/src/main/kotlin/fr/xgouchet/elmyr/junit5/ForgeConfiguration.kt
@@ -14,6 +14,6 @@ import org.junit.platform.commons.annotation.Testable
 @MustBeDocumented
 @Inherited
 annotation class ForgeConfiguration(
-    val value: KClass<out ForgeConfigurator>,
+    val value: KClass<out ForgeConfigurator> = ForgeConfigurator.NoOp::class,
     val seed: Long = 0L
 )