+++ CHANGELOG.md
@@ -1,14 +1,20 @@
 # Changelog
 
-### next version (????/??/??)
+### 1.0.0-beta3 (2019/12/09)
 
 #### `core`
 
  - Add the nullable value forgeries
+ - Add the enum forgeries
+
+#### `junit4`
+
+ - Handle enums fields annotated with `@Forgery`
 
 #### `junit5`
 
  - Fix the error message when a test fails
+ - Handle enums fields and parameters annotated with `@Forgery`
  
 #### `jvm`